requirements.txt
streamlit
audio-recorder-streamlit
openai-whisper
google-cloud-aiplatform
google-cloud-logging
